requirements.txt
cffi==1.14.6
pycparser==2.20
pycryptodome==3.10.1
PyNaCl==1.4.0
six==1.16.0
